0.6.0 snapshot

+++ restart-all.sh
@@ -5,17 +5,16 @@ export FEDORA_HOME=$ARCHIVE_HOME/fedora
 
 sudo service elasticsearch restart
 
-kill `ps -eaf|grep tomcat|awk '{print $2}'|head -1`
-
-
-#cd $ARCHIVE_HOME/regal-api
-#$ARCHIVE_HOME/play-2.2.3/play dist
-#cd -
-
 cd $ARCHIVE_HOME/regal-server
+if [ -f RUNNING_PID ]
+then
 kill `cat RUNNING_PID`
+fi
 nohup $ARCHIVE_HOME/regal-server/bin/regal-api &
 cd -
 
+kill `ps -eaf|grep tomcat|awk '{print $2}'|head -1`
+sleep 10
 $TOMCAT_HOME/bin/startup.sh
+sleep 10
